(windows) MSI install from url
MSIs can be installed from a URL so there is not a need to download it
separately. This also switches the order of arguments passed to msiexec to be
more appropriate.

This also allows the MSI location to be local or on a share since it won't
attempt to download it first.